Creates indexes that would have been created if we were able to specify foreign keys (cr-sqlite doesn't allow them).

Current benchmarks:

Route: /stock:
  - getPublisherList: 125 ms
  - getStock: 168 ms
  - **total load: 126 ms**

Route: /inventory/warehouses:
  - getAllWarehouses: 533 ms
  - **total load: 533 ms**

Route: /inventory/warehouses/[id]:
  - getWarehouseById: 2 ms
  - getStock: 431 ms
  - getPublisherList: 150 ms
  - **total load: 583 ms**

Route: /inventory/inbound:
  - getActiveInboundNotes: 1157 ms _(note: with all notes set to committed = 0, probably won't be as many in production)_
  - **total load: 1158 ms**

Route: /inventory/inbound/[id]:
  - getNoteById: 3 ms
  - getNoteEntries: 3 ms
  - getPublisherList: 146 ms
  - **total load: 152 ms**

Route: /outbound:
  - getActiveOutboundNotes: 774 ms _(note: with all notes set to committed = 0, probably won't be as many in production)_
  - **total load: 774 ms**

Route: /outbound/[id]:
  - getNoteById: 2 ms
  - getAllWarehouses: 372 ms
  - getNoteEntries: 3 ms
  - getStock: 5 ms
  - getNoteCustomItems: 1 ms
  - getPublisherList: 134 ms
  - **total load: 518 ms**

Route: /history/date:
  - getPastTransactions: 824 ms
  - **total load: 845 ms**

Route: /history/date/[date]:
  - getPastTransactions: 753 ms
  - **total load: 754 ms**

Route: /history/isbn:
  - searchBooks: 136 ms

Route: /history/isbn/[isbn]:
  - getPastTransactions: 5 ms
  - getBookData: 1 ms
  - getStock: 2 ms
  - **total load: 8 ms**

Route: /history/notes:
  - getPastNotes: 30 ms
  - **total load: 30 ms**

Route: /history/notes/[date]:
  - getPastNotes: 8 ms
  - **total load: 9 ms**

Route: /history/notes/archive/[id]:
  - getNoteById: 2 ms
  - getNoteEntries: 1 ms
  - getNoteCustomItems: 1 ms
  - **total load: 5 ms**

Route: /history/warehouse:
  - getAllWarehouses: 481 ms
  - **total load: 481 ms**

Route: /history/warehouse/[warehouseId]:
  - getWarehouseById: 2 ms
  - getPastTransactions: 130 ms
  - **total load: 132 ms**

Route: /history/warehouse/[warehouseId]/[from]/[to]/[[noteType]]:
  - getWarehouseById: 3 ms
  - getPastTransactions: 233 ms
  - **total load: 236 ms**